WDT-257 - Don't create or update security configuration adjudicators. Warn if the model has non-default configuration.
