+++ core/src/main/resources/oracle/weblogic/deploy/messages/wlsdeploy_rb.properties
@@ -980,6 +980,12 @@ WLSDPLY-12134=Unable to remove "{0}" in target domain release {1}. The remove is
   the Realm Provider Type {2}. Consult the WebLogic Deploy Tool documentation for further information. : {3}
 WLSDPLY-12135=Removing Security provider {0} with Provider type {1} at location {2}
 WLSDPLY-12136=No default providers installed for {0} at {1}
+WLSDPLY-12137=Update of security configuration adjudicators is not supported
+WLSDPLY-12138=Security configuration adjudicator count is {0}
+WLSDPLY-12139=Security configuration adjudicator name is {0}
+WLSDPLY-12140=Security configuration adjudicator has {0} subtypes
+WLSDPLY-12141=Security configuration adjudicator subtype is {0}
+WLSDPLY-12142=Security configuration adjudicator has {0} attributes
 
 # domain_creator.py
 WLSDPLY-12200={0} did not find the required {1} section in the model file {2}
